feat(cli): login to remote gitlab by default
